Move handling tickets to seperate controller and fix event UpdateDisplayAboutTicket
